pyproject.toml
[project]
name = "claude-p"
version = "0.1.0"
description = "Home server for Claude Code agent jobs"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Hayk Harutyunyan" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "pyyaml>=6.0",
    "watchfiles>=0.24",
    "croniter>=3.0",
    "wsgidav>=4.3",
    "a2wsgi>=1.10",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "argon2-cffi>=23.1",
    "itsdangerous>=2.2",
    "sse-starlette>=2.1",
    "httpx>=0.27",
]

[project.scripts]
claude-p = "claude_p.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_p"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
